requirements.txt
numpy==2.0.2
torchvision==0.23.0
gym==0.26.2
matplotlib